generative-text/requirements.txt
Flask
python-dotenv
groq